# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto | WASA Project 2022/2023
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks
    to WASAPhoto! You can upload your photos directly from your PC, and they
    will be visible to everyone following you.
  version: "1"
  contact:
    name: "Antonio Pietro Romito - mat. 1932500"
    email: "romito.1932500@studenti.uniroma1.it"
    url: "https://github.com/Romitoskj"
servers:
- url: http://localhost:3000
...
paths:
...
  /users/:
    parameters:
      - name: search
        in: query
        required: true
        description: The searched username.
        schema: { $ref: '#/components/schemas/name' }
    get:
      security:
        - bearerAuth: []
      tags:
        - User Profiles
      operationId: searchUser
      summary: Search an user by username
      description: |-
        Return a list of users whose username match the given query string.
      responses:
        '200': { $ref: '#/components/responses/200SearchUser' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
  /users/{user}:
    parameters:
      - { $ref: '#/components/parameters/user' }
    get:
      security:
        - bearerAuth: []
      tags:
        - User Profiles
      operationId: getUserProfile
      summary: Get the user profile page
      description: |-
        Return the user profile page of the specified user with its information
        (id, username, number of followers, number of following and number of
        photos uploaded) and photos in revers chronological order.
      responses:
        '200': { $ref: '#/components/responses/200GetUserProfile' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
  /users/{user}/username:
    parameters:
      - { $ref: '#/components/parameters/user' }
    put:
      security:
        - bearerAuth: []
      tags:
        - User Profiles
      operationId: setMyUserName
      summary: Update the username
      description: |-
        Update the username of the given user with the one provided in the
        request body.
      requestBody:
        description: The new username.
        content:
          application/json:
            schema: { $ref: '#/components/schemas/username' }
      responses:
        '204': { $ref: '#/components/responses/204Update' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
...
  /users/{user}/photos/:
    parameters:
      - { $ref: '#/components/parameters/user' }
    post:
      security:
        - bearerAuth: []
      tags:
        - Photos
      operationId: uploadPhoto
      summary: Upload a photo
      description: Upload a photo on the profile of the provided user.
      requestBody:
        description: The image file to upload.
        content:
          image/jpeg:
            schema: { $ref: '#/components/schemas/image' }
      responses:
        '201': { $ref: '#/components/responses/201Photo' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
  /users/{user}/photos/{photo}:
    parameters:
      - { $ref: '#/components/parameters/user' }
      - { $ref: '#/components/parameters/photo' }
    get:
      security:
        - bearerAuth: []
      tags:
        - Photos
      operationId: getPhoto
      summary: Get a specific photo of an user
      description: |-
        Return the jpeg file of the specified photo of the provided user.
      responses:
        '200': { $ref: '#/components/responses/200Photo' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }

    delete:
      security:
        - bearerAuth: []
      tags:
        - Photos
      operationId: deletePhoto
      summary: Delete a photo
      description: |-
        Delete the specified photo of the provided user. Removal of a photo also
        removes likes and comments.
      responses:
        '204': { $ref: '#/components/responses/204Delete' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    name:
      title: Name
      type: string
      example: Maria
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      description: This string represents the effective username.
    username:
      title: Username
      description: This object represents the username chosen by an user.
      type: object
      properties:
        name: { $ref: '#/components/schemas/name' }
    id:
      title: ID
      type: integer
      example: 123
      description: An integer auto-incremental ID.
...
    user:
      title: User
      description: This object represents a single user.
      type: object
      properties:
        identifier: { $ref: '#/components/schemas/id' }
        name: { $ref: '#/components/schemas/name' }
    users:
      title: Users
      description: List of user objects.
      type: array
      minItems: 0
      maxItems: 1000
      items: { $ref: '#/components/schemas/user' }
    photo:
      title: Photo
      description: This object represents a single photo.
      type: object
      properties:
        identifier: { $ref: '#/components/schemas/id' }
        author: { $ref: '#/components/schemas/name' }
        created_at:
          type: string
          format: date-time
          example: "2023-08-29T09:11:30.133Z"
          description: Upload date and time.
        likes_n:
          type: integer
          example: 100
          description: Likes count.
        comments_n:
          type: integer
          example: 20
          description: Comments count.
    image:
      title: Image
      description: This object represents the image file of a photo.
      type: string
      format: binary
      minLength: 1
      maxLength: 100000000000
    photos:
      title: Photos
      description: List of photos objects in revers chronological order.
      type: array
      minItems: 0
      maxItems: 1000
      items: { $ref: '#/components/schemas/photo' }
    profile:
      title: Profile
      description: This object represents a profile page of a single user.
      type: object
      properties:
        name: { $ref: '#/components/schemas/name' }
        photos_n:
          description: Number of photos uploaded.
          type: integer
          example: 15
        followers_n:
          description: Number of follower of the user.
          type: integer
          example: 1000
        following_n:
          description: Number of accounts that the user follows.
          type: integer
          example: 1000
        photos: { $ref: '#/components/schemas/photos' }
        followed:
          description: True if the authenticated user follow the user.
          type: boolean
          example: true
        owner:
          description: True if the authenticated user is the owner of the profile.
          type: boolean
          example: false
...
    error:
      title: Error
      type: object
      description: Error message.
      properties:
        message:
          type: string
          description: The message.
          example: Error message.
  parameters:
    user:
      name: user
      in: path
      required: true
      description: The user identifier.
      schema: { $ref: '#/components/schemas/id' }
...
    photo:
      name: photo
      in: path
      required: true
      description: The photo identifier.
      schema: { $ref: '#/components/schemas/id' }
...
  responses:
    200SearchUser:
      description: List of users.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/users' }
    200GetUserProfile:
      description: User profile page.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/profile' }
...
    200Photo:
      description: A single photo.
      content:
        image/jpeg:
          schema: { $ref: '#/components/schemas/image' }
...
    201Photo:
      description: The photo has been uploaded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/photo' }
...
    204Update:
      description: The resource has been updated.
    204Delete:
      description: The resource has been deleted.
    400BadRequest:
      description: The request is not valid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error'}
          example: {"message": "The request body is not valid."}
    401Unauthorized:
      description: Authentication ID is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error'}
          example: {"message": "Authentication ID is missing or invalid."}
    403Forbidden:
      description: |-
        Action proibithed when performing an operation without permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error'}
          example: {"message": "This resource belongs to someone else."}
    404NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error'}
          example: {"message": "Resource not found."}
    500InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error'}
          example: {"message": "Internal server error."}